+++ hivemind/__init__.py
@@ -1,7 +1,6 @@
 from hivemind.client import *
 from hivemind.dht import *
-from hivemind.server import Server
+from hivemind.server import *
 from hivemind.utils import *
-from hivemind.runtime import *
 
 __version__ = '0.7.1'
